Fix a problem with arbitrary, time seeks.
Commit afaeef4a mistakenly changed a case that really shold use
integer division.

+++ mythtv/libs/libmythtv/tv_play.cpp
@@ -6634,7 +6634,7 @@ void TV::DoArbSeek(PlayerContext *ctx, ArbSeekWhence whence,
     if (!ok)
         return;
 
-    float time = ((seek / 100.0F) * 3600) + ((seek % 100) * 60);
+    float time = (int(seek / 100) * 3600) + ((seek % 100) * 60);
 
     if (whence == ARBSEEK_FORWARD)
         DoSeek(ctx, time, tr("Jump Ahead"),
